requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bracex==2.3.post1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-braces==1.15.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-taggit==3.1.0
django_debug_toolbar==3.8.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fqdn==1.5.1
ghp-import==2.1.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.4.2
mkdocs-awesome-pages-plugin==2.8.0
mkdocs-material==9.0.5
mkdocs-material-extensions==1.1.1
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
oauthlib==3.2.2
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python3-openid==3.2.0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
